requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
fonttools==4.47.2
idna==3.6
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.2
newsapi-python==0.2.7
nltk==3.8.1
numpy==1.26.3
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
soupsieve==2.5
textblob==0.17.1
tqdm==4.66.1
tzdata==2023.4
urllib3==2.2.0
